上传更改密码，首页商品接口跳转详情页，详情页跳转店铺首页接口实现。

+++ app/src/main/java/com/example/administrator/xiangou/mine/mystore/couponmanager/CouponManagerContract.java
@@ -19,6 +19,6 @@ public class CouponManagerContract {
 
     interface  Presenter extends BasePresenter<View> {
         void callFindCoupon(int did,String condition);
-        void callAddCoupon(int did,double condition,double money,int createnum,int use_start_time,int use_end_time);
+        void callAddCoupon(int did,double condition,double money,int createnum,long use_start_time,long use_end_time);
     }
 }
